Update cloudfront.md documentation
I just wanted to clarify what types of requests to `/js/script.js` and `/api/event` would result in successful responses. I spent a lot of time wondering why a GET request to `/api/event` would 404 instead of returning a 202. Sending a POST request via curl seems to be fine, and the API docs seem to indicate there isn't a GET request for `/api/event` so it doesn't seem like it's a problem for that to 404.